Stars and Warnings
Managed to reduce the skipping the stars give off when the program first starts. Also got rid of all the warnings the compiler gives off when building. May need to redo the particle system for the stars depending on performance.

+++ SFML_Asteroids/SFML_Asteroids/GameObject.cpp
@@ -37,7 +37,7 @@ void GameObject::Update(sf::RenderWindow* Window, float DeltaTime)
 	*/
 	if (m_Position.x < 0) // Left
 	{
-		m_Position.x = Window->getSize().x;
+		m_Position.x = (float)Window->getSize().x;
 		WentOffEdgeOfScreen(false);
 	}
 	else if (m_Position.x > Window->getSize().x) // Right
@@ -48,7 +48,7 @@ void GameObject::Update(sf::RenderWindow* Window, float DeltaTime)
 
 	if (m_Position.y < 0) // Top
 	{
-		m_Position.y = Window->getSize().y;
+		m_Position.y = (float)Window->getSize().y;
 		WentOffEdgeOfScreen(true);
 	}
 	else if (m_Position.y > Window->getSize().y) // Bottom
